Revert "add read/write data services"
