--- src/main/roboy/memory/Parser.java
@@ -1,11 +1,0 @@
-package memory;
-
-/**
- * Parser
- */
-public class Parser {
-
-    public void parse(String data) {
-
-    }
-}
