+++ src/main/roboy/memory/util/Config.java
@@ -11,6 +11,6 @@ public class Config {
     //Neo4j
     public final static String NEO4J_ADDRESS = "bolt://bot.roboy.org:7687";
     public final static String NEO4J_USERNAME = "memory";
-    public final static String NEO4J_PASSWORD = "************************"; //TODO: do NOT commit password to the repo
+    public final static String NEO4J_PASSWORD = "***********************"; //TODO: do NOT commit password to the repo
 
 }
